Issue 25 change project cache dir (#30)
* Project cache is placed inside target dir
* Rules for target dir change. Instead of picking the builder name, the default is `.hdlcc`. Project cache will the be at `.hdlcc/.hdlcc.cache'
* Other general improvements

+++ hdlcc/code_checker_base.py
@@ -18,11 +18,14 @@
 
 import abc
 import os
-import shutil
 import os.path as p
+import shutil
 import logging
 import traceback
+import threading
+from multiprocessing.pool import ThreadPool
 
+# Make the serializer transparent
 try:
     import json as serializer
     def _dump(*args, **kwargs):
@@ -35,9 +38,6 @@ except ImportError:
         import pickle as serializer
 
     _dump = serializer.dump # pylint: disable=invalid-name
-
-import threading
-from multiprocessing.pool import ThreadPool
 
 import hdlcc.exceptions
 import hdlcc.builders
@@ -69,32 +69,34 @@ class HdlCodeCheckerBase(object):
         self._config = None
         self.builder = None
 
-        self.buildByDependency()
+        self.setupEnvIfNeeded()
+        #  self.buildByDependency()
 
-    @staticmethod
-    def _getCacheFilename(project_file):
+    def _getCacheFilename(self, target_dir=None):
         "Returns the cache file name for a given project file"
-        return p.join(p.dirname(project_file), \
-            '.' + p.basename(project_file))
-
-    @staticmethod
-    def cleanProjectCache(project_file):
-        "Clean up cached file"
-        if project_file is None:
-            _logger.debug("Project file is None, can't clean")
-            return
-        cache_fname = HdlCodeCheckerBase._getCacheFilename(project_file)
-        if p.exists(cache_fname):
-            _logger.debug("Removing cached info in '%s'", cache_fname)
-            os.remove(cache_fname)
+        if target_dir is None:
+            if self._config is None:
+                return None
+            else:
+                target_dir = self._config.getTargetDir()
+        return p.join(target_dir, '.hdlcc.cache')
 
     def clean(self):
         "Clean up generated files"
-        self.cleanProjectCache(self._config.filename)
+        cache_fname = self._getCacheFilename()
+        if cache_fname is not None and p.exists(cache_fname):
+            _logger.debug("Removing cached info in '%s'", cache_fname)
+            os.remove(cache_fname)
+
         target_dir = self._config.getTargetDir()
         if p.exists(target_dir):
             _logger.debug("Removing target dir '%s'", target_dir)
             shutil.rmtree(target_dir)
+
+        del self._config
+        del self.builder
+        self._config = None
+        self.builder = None
 
     def _setState(self, state):
         "serializer load implementation"
@@ -179,8 +181,9 @@ class HdlCodeCheckerBase(object):
                 yield source
 
             if empty_step:
-                for missing_path in \
-                        list(set(self._config.getSourcesPaths()) - set(sources_built)):
+                missing_paths = list(set(
+                    self._config.getSourcesPaths()) - set(sources_built))
+                for missing_path in missing_paths: # pragma: no cover
                     source = self._config.getSourceByPath(missing_path)
                     dependencies = self._getSourceDependenciesSet(source)
                     missing_dependencies = dependencies - set(self._units_built)
@@ -206,8 +209,10 @@ class HdlCodeCheckerBase(object):
                 (x['error_type'], x['line_number'], x['error_number']))
 
     def _getMessagesAvailable(self, path, *args, **kwargs):
-        '''Checks if the builder can be called via self._getBuilderMessages
-        or return info/messages identifying why it couldn't be done'''
+        """
+        Checks if the builder can be called via self._getBuilderMessages
+        or return info/messages identifying why it couldn"t be done
+        """
 
         if not self.finishedBuilding():
             self._handleUiWarning("Project hasn't finished building, try again "
@@ -288,7 +293,7 @@ class HdlCodeCheckerBase(object):
 
     def saveCache(self):
         "Dumps project object to a file to recover its state later"
-        cache_fname = self._getCacheFilename(self._config.filename)
+        cache_fname = self._getCacheFilename()
 
         state = {'serializer' : serializer.__name__,
                  '_logger': {'name' : self._logger.name,
@@ -300,14 +305,18 @@ class HdlCodeCheckerBase(object):
         self._logger.debug("Saving state to '%s'", cache_fname)
         _dump(state, open(cache_fname, 'w'))
 
-    def _recoverCache(self):
-        '''Tries to recover cached info for the given project_file.
-        If something goes wrong, assume the cache is invalid and return
-        nothing. Otherwise, return the cached object'''
-        if self.project_file is None:
-            self._logger.debug("Can't recover cache from None")
+    def _recoverCache(self, target_dir):
+        """
+        Tries to recover cached info for the given project_file. If
+        something goes wrong, assume the cache is invalid and return
+        nothing. Otherwise, return the cached object
+        """
+        cache_fname = self._getCacheFilename(target_dir)
+        #  if self.project_file is None or cache_fname is None:
+        if cache_fname is None:
+            self._logger.warning("Can't recover cache from None")
             return
-        cache_fname = self._getCacheFilename(self.project_file)
+
         _logger.debug("Trying to recover from '%s'", cache_fname)
         cache = None
         if p.exists(cache_fname):
@@ -359,18 +368,20 @@ class HdlCodeCheckerBase(object):
         with self._lock:
             self._logger.info("Build has finished")
 
-    def _setupEnvIfNeeded(self):
-        '''Updates or creates the environment, which includes checking
+    def setupEnvIfNeeded(self):
+        """
+        Updates or creates the environment, which includes checking
         if the configuration file should be parsed and creating the
-        appropriate builder objects'''
-
-        _logger.debug("Current builder is %s", self.builder)
-
+        appropriate builder objects
+        """
         try:
-            # No builder defined means it's the first time we run, so
-            # try to recover from cache
-            if self.builder is None:
-                self._recoverCache()
+            # If the configuration is undefined, try to extract the
+            # target dir from the project file so we can have a hint of
+            # where the cache file should be
+            if self._config is None and self.project_file is not None:
+                target_dir, _ = ConfigParser.simpleParse(self.project_file)
+                if target_dir:
+                    self._recoverCache(target_dir)
 
             # No configuration defined means we failed to recover it
             # from the cache
@@ -397,7 +408,7 @@ class HdlCodeCheckerBase(object):
     def _buildByDependency(self):
         "Build the project by checking source file dependencies"
         with self._lock:
-            self._setupEnvIfNeeded()
+            self.setupEnvIfNeeded()
             built = 0
             errors = 0
             warnings = 0
@@ -421,10 +432,13 @@ class HdlCodeCheckerBase(object):
                     built, errors, warnings)
 
     def getMessagesByPath(self, path, *args, **kwargs):
-        '''Returns the messages for the given path, including messages
+        """
+        Returns the messages for the given path, including messages
         from the import configured builder (if available) and static
-        checks'''
-        if not p.isabs(path):
+        checks
+        """
+        self.setupEnvIfNeeded()
+        if not p.isabs(path): # pragma: no cover
             abspath = p.join(self._start_dir, path)
         else:
             abspath = path
